CLIENT: should fix issue #107 (MAX_STREAM_BUFFER warnings)
the ogm video is using a lot of buffers - but they are all cleaned up and nothing is leaked.
This just increases the buffer count to remove the warning and the sound stutters.

It worked fine with 128 - but in valgrind I've seen them exceeding again.

The usual sound driver has 16384 - I suppose bumping it to 512 is safe.