GAME: Impact marks sometimes doesn't draw at all.
Not easy to reproduce if you don't play (io)Quake3 every day and know the places where it happens! ;)
But anyway...

Test: start q3dm12 go to "Long Jump Canyon" (where the small platform
teleporter for the BFG is) place yourself at the point where the railgun
spawns, look in the direction where the red suspended armor is. Now shoot
at the sloped wall on the out/leftside of the door you see. (the sloped
wall should be nearly in the center of your screen now). If you choose the
correct brush face and shoot up and down at this brush face, the impact
marks sometimes aren't visible.

There are hundreds of custom maps where this can happen!

eb9fe030c47281e6145908e0331cc543b4aaf432 ioq3