CGAME: Restore SB showing personal scores in team based games
With WoP 1.5, individual scores in team-based games were registered but not shown in the scoreboard, in order to not put too much emphasis on individual scores and to support the team idea more. However, this was often criticized and voices from the community would like to see the display of individual points back. Therefore, the withdrawal of this change.