CGAME: Support for player model sounds depending on the skin
This makes it possible to include a separate sound set for each player skin if desired or simply to "replace" individual sound files of a player model depending on the selected skin. The system will now first search for a folder with the same name in the sounds/wop_players/ directory using the skin name and load the sound set from there first. If this is not successful or files are missing, the folder associated with the model name of the skin is used. Only if the loading of from a folder of a skin or model fails, the default model is used, as before. The approach to create folders for each skin parallel to the models and not to stack these folders further is due to the simpler implementation. For testing, PadBabe can now be loaded and has a different taunt sound than PadGirl. The sound file is in the directory since the PadMod, but was never used. See #18 for further information.