WOP: Reorganization of sound files (step 1)
The sound files have been evolutionarily placed in two directories /sound and /sounds. The former is natively supported by the level editors, the latter is not. The default directory for idTech3 engine sound files is /sound. This is a first step to reorganize the sound files and eliminate the /sounds directory. This represents a break with previous WoP versions, but we don't need to worry about backwards compatibility here. Custom content can be easily adapted, even without having to recompile maps, for example.