WOP: Reorganization of sound files (step 4)
Transfer of health station sounds from /sounds to /sound. The station_still sound has never been used, but is contains a heartbeat sound. I would like to hear this as an idle sound for the health stations. If implemented in the future, the revival items heartbeat sound should be used instead sound/items/revival/heartbeat.