CGAME/GAME/WOP: Shortening sound paths
Taking the chance now of shortening sound paths, see: https://forum.clover.moe/t/question-about-configstrings-name-length/49.html. This will break the sounds used in maps and will force a recompile or hex edit of BSPs plus an AAS recompile. With map updates on the way the time for this is now.