GAME/CGAME/UI: "impressive" award cleanup
"Impressive" is not an award in WoP. WoP lets players laugh "hehe2" when two SPALSHER hits are made in a row. There is no need to count this. The "impressive" medal sprite and sound have already been disabled in the past. This cleanup gives room (16bits is not much) for other awards in the future.