UI/WOP: Removes shadows from the arrows in player settings menu
The shadows of the arrows do not combine well when they are grayed out. All other arrows in the menu have no shadows. It is better to remove them and use the extra texture space for higher resolution. This change also requires a slight adjustment in size and position coordinates in the code.