CGAME: Pumper weapon fx for freeze tag
Pumper weapon fx from the XMAS mod for freeze tag. Additionally changed the flash delight color from green to white in freeze tag.
Note: there is still a texture flashing up between weapon and user on screen while pressing the fire button which seems to be green. I couldn't find the source for it to change it to white.