XMAS: Fixes several issues with in-game model shaders and skins not working correctly
The XMAS mod shaders were created for WoP 1.2. In this version many shaders were stored in a huge number of separate shader files. With WoP 1.5, many of these shaders were combined into a few files. To make the XMAS mod shaders work again, actually these shader files would have to be copied again and all the XMAS mod changes would have to be pasted again. However, this also leads to a large redundancy of unchanged shader entries between WoP and XMAS-Mod. To avoid this, with this commit most XMAS shaders are now combined in the file wop_xmas.shader. To ensure that they still work and are not still overwritten by the WoP shaders, the paths within the shaders and their associated models have been changed. In some cases, textures have also been renamed for better differentiation. Thus, the function of the XMAS shaders is fully restored and the design of the XMAS mod remains unaffected.