GAME/CGAME: Turns g_enableDust into a client side variable (cg_enableDust)
Due to this change, it is not necessary to activate the dust feature via the worldspawn of the map any more. It should be sufficient now to enable it by setting the surfaceparm dust for the desired surface via a shader. The CVAR g_enableDust is now changed to a client side variable cg_enableDust with default value 1 to have a switch for the feature on the client side.