UI: Subtracts an offset to the height in Help menus for non-widescreen resolutions
For non-widescreen resolutions, we need to calculate and subtract an offset from the height so that the background graphics and arrows always have the same distance to the top of the screen. Also, UI_DrawHelpMenuImg was merged with UI_HelpMenu_Draw because UI_DrawHelpMenuImg had only one line of code.