UI: Fixes antialiasing not set by the menu when opengl2
Fixes that antialiasing (r_ext_framebuffer_multisample) is not set via the menu. The option in the graphics menu now always sets the value for r_ext_multisample AND r_ext_framebuffer_multisample, regardless of which renderer is active. If r_ext_multisample and r_ext_framebuffer_multisample have different values, e.g. set by a console command, the value for r_ext_multisample or r_ext_framebuffer_multisample is read out and displayed in the menu, depending on the selected renderer.