CGAME: Remove cvar cg_drawIcons
We always draw all the icons so we don't need it. This worked well to tweak the Q3A HUD and scoreboard, but the WoP HUD is different and in most cases cg_drawIcons was removed from HUD code anyway. It only worked with the scoreboard but is kind of useless.