PADPACK: Seperate skybox and levelshots shader
Bad idea to combine all shaders, because the OpenGL2 sunlight key q3gl2_sun is causing the skybox not to work with OpenGL1. So we cannot combine q3gl2_sun and q3map_sun in one shader, which is a bummer. We always need to overwrite the OpenGL1 shader with the OpenGL2 shader by adding an mtr file with the same name. But in this case we need to copy all shader entries within the file which is a lot of redundant data. So it makes sense to hold only the skyboxes in a separate shader file.