PADPACK: Adjust q3gl2_sun light source parameters
PadGallery and PadCastle have their point light sources, the Moon and the Sun. The original skybox values do not correspond to these point light sources. To avoid the weird effect of light coming from the opposite direction with the OpenGL2 renderer, the values had to be adjusted. It's not perfect, but it's a compromise to look good with the baked lightmap.