WOP/PADPACK: Replace fence texture
Since the texture did not have an aspect ratio of a power of 2, it was drawn distorted, causing artifacts in the game, hence the replacement. Due to the higher resolution of the replacement, the texture had to be scaled a bit in the shaders used to make it suitable for the maps used. Overall, the quality should be much better.