WOP: Extract PadLibrary skybox to pad_skyboxes shader
Adapting the skybox shaders to work with the OpenGL2 renderer. Adjustment of intensity, degrees, and elevation was necessary to make the given point light sources of the skyboxes, such as the sun, fit and look believable. See shadows in the spray room especially when some animated brush work comes into action, simply stunning.