WOP: Extract PadShop skybox to pad_skyboxes shader
Adapting the skybox shaders to work with the OpenGL2 renderer. Adjustment of color, intensity, degrees, and elevation was necessary to make the given point light sources of the skyboxes, such as the sun, fit and look believable. The map uses two skyboxes at the same time, so I needed to disable q3map_sun of the main map to make q3gl2_sun of the mercury skybox work. pc-ground-zero512 skybox should not have a point light source anyways. It's night, no moon, so only diffuse light sources are available.