WOP: Extract Jail skyboxes to pad_skyboxes shader
Adapting the skybox shaders to work with the OpenGL2 renderer. Adjustment of intensity was necessary to make the given point light source of the skyboxes, such as the sun, fit and look believable. The map uses two skyboxes at the same time, but with same values, so no issues here.