WOP: Extract Airplane skybox to pad_skyboxes shader
Adapting the skybox shader to work with the OpenGL2 renderer. Sadly this map does not use the skylensflare feature. Also point light source (sun) on the skybox does not exactly fit the skybox shaders values. Both stays a ToDo until a recompile of this map is possible. Changing values only in gl2 shader now would lead to a big and visible mismatch with the baked lightmap.