WOP: Extract PadAttic skyboxs to pad_skyboxes shader
Adapting the skybox shaders to work with the OpenGL2 renderer. Unfortunately, the night skybox of the LPS version has a mismatch between the moon on the texture and the set point light source in the shader. The gl2 shader version corrects this to make the moon a believable light source, but now there is a mismatch between the baked shadows and the gl2 shadows. This cannot be fixed without a recompile. Also, PadAttic's morning skybox does not have a point light source on the texture, nor is the sky lensflare feature used to simulate one. So there is no way to get a believable point light with gl2 shadows.