WOP: Extract CybJourney skybox to pad_skyboxes shader
Adapting the skybox shader to work with the OpenGL2 renderer. Adding a sky lens flare to the map to mimic the sun as a point source of light according to given shadows drawn on the skybox. The skybox shader had to be adjusted in intensity, degrees and elevation. This then allowed gl2 shadows to be introduced with amazing results due to the animated map brushwork casting dynamic shadows. A recompile was also needed to adjust the baked lightmap to avoid a visible lightmap mismatch between baked and dynamic shadows.