CGAME: Remove cg_thirdPersonAutoSwitch
Removes the CVAR cg_thirdPersonAutoSwitch, which allows the user to switch from 1st person view to 3rd person view when the punchy is selected and back when switching to another weapon. It interfered with cg_thirdPerson toggle in undesirable ways and was basically not used at all, aside from the lack of documentation and menu option. This solves the #162.