RENDERER: Fix opengl1/OpenGL2 crash in levelshot command
Setting packAlign was removed [1] and later set to 0 [2] and this caused
PADP( ..., 0 ) to return NULL which later gets passed to R_GammaCorrect()
and crashed.

It seems I left packAlign in my PNG/levelshot changes [3] even though
it's not needed in R_LevelShot().

Set packAlign to 1 which is single byte alignment and doesn't allocate
additional memory. PADP( ..., 1 ) returns the same address and it
doesn't crash later.

1: b514a01b461719904182d62d57554ca7e2c5971a
2: 688f4f50dd091f6f5f5368be8576c8f8f03a99a6
3: https://github.com/ioquake/ioq3/pull/262