UI/WOP: Pulse the accept button when shown
ENTE suggests to let the accept button pulse like the exit or back buttons when it appears even if it has no focus. This should point the users attention to this button to accept changes which need a restart of the game. Added on/off versions of the button with shadow because it is used on top of a bachground picture like the exit button.