GAME: Prevent players to re-collect the KillerDuck too fast in CATCH
Extend the time that the player who dropped the KillerDuck is unable to collect it by one second beyond its respawn time. This prevents players from gaining too much of an advantage from health and armor refills, especially when there are few human players on the server. This should fix #361.